# swagger_example/swagger_config.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.0",
    "title": "Task Organizer",
    "description": "Organize and maintain tasks"
  },
  "host": "127.0.0.1:5000",
  "schemes": [
    "http"
  ],
  "paths": {
    "/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Returns a list of tasks.",
        "description": "Optional extended description in Markdown.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "405": {
            "description": "Invalid Input"
          }
        }
      }
    },
    "/task/{taskname}": {
      "get": {
        "tags": [
          "Task specific activity"
        ],
        "summary": "Returns a task by name.",
        "parameters": [
          {
            "in": "path",
            "name": "taskname",
            "required": true,
            "type": "string",
            "minimum": 1,
            "description": "Name of the task to be obtained"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched task"
          }
        }
      },
      "delete": {
        "tags": [
          "Task specific activity"
        ],
        "summary": "Delete a task by name.",
        "parameters": [
          {
            "in": "path",
            "name": "taskname",
            "required": true,
            "type": "string",
            "minimum": 1,
            "description": "Name of the task to be deleted"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched task"
          }
        }
      }
    },
    "/task": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Add a task",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "name",
            "type": "string",
            "required": true,
            "description": "Name of the task"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "required": true,
            "description": "Description of the task"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}